requirements.txt
flask>=2.0.0
rich>=10.0.0
PyPDF2>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
scikit-learn>=1.0.0
numpy>=1.21.0 